app/requirements.txt
## ___ START
#Django==1.9.2
#django-sitetree==1.5.1
#requests==2.9.1
#rpyc==3.3.0

## Na toevoegen Capcha:	http://django-simple-captcha.readthedocs.io/en/latest/usage.html#installation
#Django==1.9.2
#django-ranged-response==0.2.0
#django-simple-captcha==0.5.6
#django-sitetree==1.5.1
#Pillow==5.0.0
#plumbum==1.6.3
#requests==2.9.1
#rpyc==3.3.0
#six==1.11.0

## Na toevoegen CMS:   http://docs.django-cms.org/en/latest/how_to/install.html
#dj-database-url==0.5.0
#Django==1.9.2
#django-classy-tags==0.8.0
#django-cms==3.5.1
#django-filer==1.3.0
#django-formtools==2.1
#django-mptt==0.8.7
#django-polymorphic==1.3
#django-ranged-response==0.2.0
#django-sekizai==0.10.0
#django-simple-captcha==0.5.6
#django-sitetree==1.5.1
#django-treebeard==4.2.1
#djangocms-admin-style==1.2.7
#djangocms-attributes-field==0.3.0
#djangocms-column==1.7.0
#djangocms-file==2.0.2
#djangocms-googlemap==1.1.1
#djangocms-link==2.1.2
#djangocms-picture==2.0.6
#djangocms-snippet==2.0.0
#djangocms-style==2.0.2
#djangocms-text-ckeditor==3.5.3
#djangocms-video==2.0.4
#easy-thumbnails==2.4.2
#html5lib==0.9999999
#Pillow==5.0.0
#plumbum==1.6.3
#pytz==2018.3
#requests==2.9.1
#rpyc==3.3.0
#six==1.11.0
#tzlocal==1.5.1
#Unidecode==0.4.21

## na toevoegen  mathjax   https://pypi.python.org/pypi/django-mathjax/0.0.5
dj-database-url==0.5.0
Django==1.9.2
django-classy-tags==0.8.0
django-cms==3.5.1
django-filer==1.3.0
django-formtools==2.1
django-mathjax==0.0.8
django-mptt==0.8.7
django-polymorphic==1.3
django-ranged-response==0.2.0
django-sekizai==0.10.0
django-simple-captcha==0.5.6
django-sitetree==1.5.1
django-treebeard==4.2.1
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-link==2.1.2
djangocms-picture==2.0.6
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.3
djangocms-video==2.0.4
easy-thumbnails==2.4.2
html5lib==0.9999999
Pillow==5.0.0
plumbum==1.6.3
pytz==2018.3
requests==2.9.1
rpyc==3.3.0
six==1.11.0
tzlocal==1.5.1
Unidecode==0.4.21

 

